+++ Android/UdacityExercises/Lesson10-Hydration-Reminder/T10.05-Exercise-ChargingBroadcastReceiver/app/src/main/java/com/example/android/background/MainActivity.java
@@ -20,6 +20,8 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.os.BatteryManager;
+import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;