+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 24.09.2018
+
+* [Android] Sunshine: Synchronizing The Weather
+
 #### 23.09.2018
 
 * [Android] Hydration Reminder: Getting Battery State